requirements.txt
fastapi
pydantic
faiss-cpu
PyPDF2
python-docx
uvicorn
python-multipart
chromadb
sentence-transformers
langchain
langchain-community
docx2txt
pillow
torch
transformers
pypdf
python-jose
httpx
python-dotenv